website/requirements.txt
mysql-connector==2.2.9
numpy==1.24.1
pandas==1.5.2
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
